Errors Handled

+++ Assignment3/Test/BPTree.java
@@ -200,8 +200,22 @@ public class BPTree implements Serializable{
             {
                 System.out.print("Enter the key: ");
                 int key = Reader.nextInt();
-                int out = bt.search(key);
-                System.out.println(bt.data.get(out));
+                boolean flag = false;
+                for (int i = 0; i < bt.data.size(); i++) {
+                    if (bt.data.get(i).getValidityTag().compareTo("0000") != 0) {
+                        int k = Integer.valueOf(bt.data.get(i).getInstructorId());
+                        if (k == key) {
+                            flag = true;
+                            break;
+                        }
+                    }
+                }
+                if (flag) {
+                    int out = bt.search(key);
+                    System.out.println(bt.data.get(out));
+                } else {
+                    System.out.println("Wrong key number!\nTry Again");
+                }
             }
             else if ( input == 2 )
             {
@@ -229,8 +243,22 @@ public class BPTree implements Serializable{
                 // Since Keys are unique there will be only one data.
                 System.out.print("Enter the key for which you want to print all data: ");
                 int key = Reader.nextInt();
-                int out = bt.search(key);
-                System.out.println(bt.data.get(out));
+                boolean flag = false;
+                for (int i = 0; i < bt.data.size(); i++) {
+                    if (bt.data.get(i).getValidityTag().compareTo("0000") != 0) {
+                        int k = Integer.valueOf(bt.data.get(i).getInstructorId());
+                        if (k == key) {
+                            flag = true;
+                            break;
+                        }
+                    }
+                }
+                if (flag) {
+                    int out = bt.search(key);
+                    System.out.println(bt.data.get(out));
+                } else {
+                    System.out.println("Wrong key number!\nTry Again");
+                }
             }
             else if ( input == 5 )
             {
@@ -243,33 +271,55 @@ public class BPTree implements Serializable{
                 String department = Reader.next();
                 System.out.print("Enter Salary: ");
                 int salary = Reader.nextInt();
-                // Write to the file and insert in Tree
-                Data obj = new Data(id, name, department, salary);
-                bt.data.add(obj);
-                bt.insert(Integer.valueOf(id), bt.data.size()-1);
-                createCSV(bt.data, "local_record.csv");
-                File file = new File("indexfile");
+                boolean Flag = true;
+                for (int i = 0; i < bt.data.size(); i++) {
+                    int inID = Integer.valueOf(bt.data.get(i).getInstructorId());
+                    if (inID == Integer.valueOf(id) && bt.data.get(i).getValidityTag().compareTo("0000") != 0) {
+                        Flag = false;
+                        break;
+                    }
+                }
+                if (Flag) {
+                    // Write to the file and insert in Tree
+                    Data obj = new Data(id, name, department, salary);
+                    bt.data.add(obj);
+                    bt.insert(Integer.valueOf(id), bt.data.size() - 1);
+                    createCSV(bt.data, "local_record.csv");
+                    File file = new File("indexfile");
 
-                file.delete();
-                serialize(bt.data);
-                createCSV(bt.data, "local_record.csv");
+                    file.delete();
+                    serialize(bt.data);
+                    createCSV(bt.data, "local_record.csv");
+                } else {
+                    System.out.println("Record already exist with same InstructorID.");
+                }
             }
             else if ( input == 6 )
             {
                 System.out.print("Enter the key: ");
                 int key = Reader.nextInt();
-                bt.delete(key);
-                for ( int i=0; i<bt.data.size(); i++ )
-                {
-                    Data e = bt.data.get(i);
-                    if ( Integer.valueOf(e.getInstructorId()) == key )
-                    {
-                        e.Delete();
+                boolean Flag = false;
+                for (int i = 0; i < bt.data.size(); i++) {
+                    int inID = Integer.valueOf(bt.data.get(i).getInstructorId());
+                    if (inID == key && bt.data.get(i).getValidityTag().compareTo("0000") != 0) {
+                        Flag = true;
                         break;
                     }
                 }
-                serialize(bt.data);
-                createCSV(bt.data, "local_record.csv");
+                if (Flag) {
+                    bt.delete(key);
+                    for (int i = 0; i < bt.data.size(); i++) {
+                        Data e = bt.data.get(i);
+                        if (Integer.valueOf(e.getInstructorId()) == key) {
+                            e.Delete();
+                            break;
+                        }
+                    }
+                    serialize(bt.data);
+                    createCSV(bt.data, "local_record.csv");
+                } else {
+                    System.out.println("Wrong key Entered.");
+                }
             }
             else
             {
